More progress on set Macros Dialog. Other minor fixes

+++ app/src/main/java/com/untrustworthypillars/simplefoodlogger/FoodManager.java
@@ -44,7 +44,7 @@ public class FoodManager {
         mCommonFoodDatabase = new CommonFoodDbHelper(mContext).getWritableDatabase();
         mExtendedFoodDatabase = new ExtendedFoodDbHelper(mContext).getWritableDatabase();
         mPreferences = androidx.preference.PreferenceManager.getDefaultSharedPreferences(context);
-        mRecentFoodsLength = (int) Float.parseFloat(mPreferences.getString("pref_recent_foods_size", "10"));
+        mRecentFoodsLength = Integer.parseInt(mPreferences.getString("pref_recent_foods_size", "10"));
     }
 
     //Add a custom food to CustomFoodDatabase
